My-app-V1/test-app/backend/requirements.txt

fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
